Separated out txctl functionality
